Single Member bugfix
Found a bug in my code addition that caused a failure in the case where the group only had a single member. i.e. line 642 would cause a bug when the $member var only had a single entry, causing it to just be a custom psobject and not an array, therefore it didn't have an op_addition method.